% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumr.R
\name{sumr}
\alias{sumr}
\title{Summarize Numeric Variables}
\usage{
sumr(data, ...)
}
\arguments{
\item{data}{A data frame to be summarized.}

\item{...}{Optional. Specific variables to summarize. If none are specified, all numeric variables are summarized.}
}
\value{
A tibble with summarized statistics for numeric variables.
}
\description{
Summarizes numeric variables in a data frame using the \code{skimr} package.
}
\examples{
library(dplyr)
library(skimr)
library(tibble)
library(gapminder)

}
