% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumr_group_by.R
\name{sumr_group_by}
\alias{sumr_group_by}
\title{Summarize Numeric Variables by Group}
\usage{
sumr_group_by(data, by, ...)
}
\arguments{
\item{data}{A data frame to be summarized.}

\item{by}{A variable to group by.}

\item{...}{Optional. Specific variables to summarize. If none are specified, all numeric variables are summarized.}
}
\value{
A tibble with summarized statistics for numeric variables by group.
}
\description{
Summarizes numeric variables in a data frame by a specified grouping variable using the \code{skimr} package.
}
\examples{
library(dplyr)
library(skimr)
library(tibble)
library(gapminder)

# Using the sumr_group_by function for specific variables
sumr_group_by(gapminder, by = continent, lifeExp, gdpPercap)

# Using the sumr_group_by function for all numeric variables
sumr_group_by(gapminder, by = continent)
}
