% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumr_na.R
\name{sumr_na}
\alias{sumr_na}
\title{Summarize Missing Values and Unique Factors}
\usage{
sumr_na(data, ...)
}
\arguments{
\item{data}{A data frame to be summarized.}

\item{...}{Optional. Specific variables to summarize. If none are specified, all variables are summarized.}
}
\value{
A tibble with summarized statistics for missing values and unique factor levels.
}
\description{
Summarizes missing values and unique factor levels for specified variables in a data frame using the \code{skimr} package.
}
\examples{
library(dplyr)
library(skimr)
library(tibble)
library(gapminder)

# Creating a gapminder dataset with some missing values
gapminder_na <- gapminder
gapminder_na[sample(1:nrow(gapminder_na), 100), "lifeExp"] <- NA

# Using the sumr_na function for specific variables
sumr_na(gapminder_na, lifeExp, gdpPercap)

# Using the sumr_na function for all variables
sumr_na(gapminder_na)
}
