% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{tab}
\alias{tab}
\title{Create a Summary Table}
\usage{
tab(data, ...)
}
\arguments{
\item{data}{A data frame to be summarized.}

\item{...}{Columns to group by.}
}
\value{
A data frame with row totals and formatted percentages.
}
\description{
Summarizes the data by the specified columns, adds row totals, and formats
the percentages using the \code{janitor} package.
}
\examples{
library(dplyr)
library(janitor)
library(gapminder)

# Example with the gapminder dataset
gapminder \%>\% tab(continent)
}
